edit paste page
