+++ public/scripts/data/editor.js
@@ -1,0 +1,13 @@
+export class PastyEditor
+{
+    constructor()
+    {
+        this.index = 0;
+        
+        this.rootElement = null;
+
+        this.titleInput = null;
+        this.languageDropdown = null;
+        this.editor = null;
+    }
+}
