+++ public/scripts/helpers/time.js
