+++ public/scripts/main.js
@@ -1,0 +1,1 @@
+import "./components/navigation.js";
