--- public/style/home.css
@@ -1,223 +1,0 @@
-#home .paste-options
-{
-    margin-bottom: 20px;
-}
-
-#home .paste-options input
-{
-    width: 100%;
-    box-sizing: border-box;
-    background-color: var(--color-nanolight);
-    color: var(--color-white);
-    border-radius: var(--border-radius);
-    border: none;
-    font-size: var(--font-size-normal);
-    padding: 0.5rem 1rem;
-    margin: 0;
-    margin-bottom: 20px;
-    font-family: var(--font-stack);
-}
-
-#home .paste-options #expires-in-dropdown
-{
-    display: flex;
-    width: 100%;
-    justify-content: space-between;
-    background-color: var(--color-nanolight);
-    border-radius: var(--border-radius);
-}
-
-#home .paste-options #expires-in-dropdown .label
-{
-    display: inline-block;
-    background-color: var(--color-nanolight);
-    padding: 0.5rem 1rem;
-    border-radius: var(--border-radius);
-}
-
-@media screen and (min-width: 640px)
-{
-    #home .paste-options
-    {
-        display: flex;
-    }
-
-    #home .paste-options input
-    {
-        margin-bottom: 0;
-        border-top-right-radius: 0;
-        border-bottom-right-radius: 0;
-    }
-
-    #home .paste-options #expires-in-dropdown
-    {
-        justify-content: initial;
-        width: auto;
-        flex-shrink: 0;
-        border-top-left-radius: 0;
-        border-bottom-left-radius: 0;
-    }
-}
-
-#home .pasty-editor
-{
-    margin-bottom: 20px;
-}
-
-#home .pasty-editor .pasty-editor-options
-{
-    border-bottom: 2px solid var(--color-mystge);
-    display: flex;
-    flex-wrap: wrap;
-}
-
-#home .pasty-editor .pasty-editor-options .pasty-editor-delete
-{
-    padding: 0.5rem 1rem;
-    padding-right: 0;
-    background-color: var(--color-nanolight);
-    border-top-left-radius: var(--border-radius);
-    color: var(--color-mysted);
-    display: none;
-}
-
-#home .pasty-editor .pasty-editor-options .pasty-editor-delete:hover
-{
-    cursor: pointer;
-}
-
-#home .pasty-editor .pasty-editor-options .pasty-editor-title
-{
-    flex: 1 0 50%;
-    width: 100%;
-    box-sizing: border-box;
-    background-color: var(--color-nanolight);
-    color: var(--color-white);
-    border-top-right-radius: var(--border-radius);
-    border-top-left-radius: var(--border-radius);
-    border: none;
-    font-size: var(--font-size-normal);
-    padding: 0.5rem 1rem;
-    margin: 0;
-    font-family: var(--font-stack);
-}
-
-#home .pasty-editor .pasty-editor-options #language-dropdown
-{
-    flex-basis: 100%;
-    display: flex;
-    width: 100%;
-    justify-content: space-between;
-    background-color: var(--color-nanolight);
-}
-
-#home .pasty-editor .pasty-editor-options #language-dropdown .label
-{
-    display: inline-block;
-    background-color: var(--color-nanolight);
-    padding: 0.5rem 1rem;
-    border-radius: var(--border-radius);
-}
-
-@media screen and (min-width: 640px)
-{
-    #home .pasty-editor .pasty-editor-options
-    {
-        flex-wrap: initial;
-    }
-
-    #home .pasty-editor .pasty-editor-options .pasty-editor-title
-    {
-        flex: initial;
-        border-top-right-radius: 0;
-        border-bottom-right-radius: 0;
-    }
-
-    #home .pasty-editor .pasty-editor-options #language-dropdown
-    {
-        flex: initial;
-        justify-content: initial;
-        width: auto;
-        flex-shrink: 0;
-        border-top-right-radius: var(--border-radius);
-    }
-}
-
-#home .pasty-editor .CodeMirror
-{
-    height: 50vh;
-    border-radius: 0 0 var(--border-radius) var(--border-radius);
-    padding-top: 0.5rem;
-}
-
-#home .paste-options-bottom
-{
-    display: flex;
-    flex-wrap: wrap;
-    position: sticky;
-    bottom: 0;
-    background-color: var(--color-nanolight);
-    border-radius: var(--border-radius);
-    padding: 1rem;
-    z-index: 100;
-    justify-content: space-between;
-}
-
-#home .paste-options-bottom .paste-buttons
-{
-    display: flex;
-    flex-direction: column;
-}
-
-#home .paste-options-bottom .paste-buttons a
-{
-    display: block;
-}
-
-#home .paste-options-bottom .paste-buttons #create-paste
-{
-    background-color: var(--color-mystge);
-    color: var(--color-nano);
-    border-radius: var(--border-radius);
-    padding: 0.5rem 1rem;
-}
-
-#home .paste-options-bottom .paste-buttons #create-paste:hover
-{
-    background-color: var(--color-mystlue);
-}
-
-#home .paste-options-bottom .paste-options-bottom-options
-{
-    display: flex;
-    flex-direction: column;
-}
-
-#home .paste-options-bottom .paste-options-bottom-options .disabled
-{
-    color: var(--color-nanolightlight);
-}
-
-#home .paste-options-bottom-sticky
-{
-    border: 2px solid var(--color-mystge);
-    border-bottom: none;
-    border-bottom-left-radius: 0;
-    border-bottom-right-radius: 0;
-}
-
-#home .paste-options-bottom-1px
-{
-    height: 1px;
-    background-color: var(--color-nano);
-}
-
-#home .paste-options-bottom .new-pasty-editor
-{
-    flex: 1 0 50%;
-}
-
-#home #paste-create-form-hidden
-{
-    display: none;
-}
