+++ views/components/addEditorButton.dt
@@ -1,0 +1,3 @@
+a.add-editor
+    i.fa.fa-plus-square
+    p add editor
