+++ views/components/pastyEditor.dt
@@ -1,0 +1,18 @@
+.pasty-editor
+    .pasty-editor-options
+        i.fas.fa-minus-square.pasty-editor-delete
+        input.pasty-editor-title(placeholder="pasty title", type="text", autocomplete="off", spellcheck="false", value=pastyTitle)
+        .language-dropdown
+            .label language:
+            include components/dropdown
+                block options
+                    - import std.conv : to;
+                    - import pastemyst.data : languages;
+                    - foreach (ulong i, lang; languages)
+                        - bool checked = false;
+                        - if (lang["name"] == pastyLang)
+                            - checked = true;
+                        label.option
+                            input(type="radio", name="language", value=lang["name"].to!string() ~ "," ~ lang["mode"].to!string() ~ "," ~ lang["mimes"][0].to!string(), checked=checked)
+                            span= lang["name"].to!string()
+    textarea.editor(type="text", autocomplete="off", spellcheck="false")= pastyCode
